Add Config Option for a deterministic Ore Processing
In some versions of Gregtech (including GTNH iirc), ore processing does not involve chanced outputs (excluding the sifter). I've added a config option for the people preferring that option over the way involving chances. That prevents the hassle of needing to remove and add a ton of recipes via groovy or craft tweaker.